% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_load_data.R
\name{A_load_data}
\alias{A_load_data}
\title{Load Data Files}
\usage{
A_load_data()
}
\arguments{
\item{...}{Prompted text entry of .tree file name (SACCHARIS output).}
}
\value{
CAZY_TABLE_FINAL_GHxx_characterized_DATE.xlsx. Exported data table containing information from cazy.org for each sequence.
}
\description{
Loading files required for plotting annotated trees generated from SACCAHRIS v2, and sorts data frames to order them based on tree nodes.
}
\examples{
A_load_data()
test_data.json
GH16_characterized.tree
}
