% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_plots_all.R
\name{B_plots_all}
\alias{B_plots_all}
\title{Plotting}
\usage{
B_plots_all(out_dir = getwd())
}
\arguments{
\item{out_dir}{Directory to write rendered graph files to. If not given, defaults to working directory.}
}
\description{
Generate circular plots of trees generated by SACCAHRIS v2. REQUIRED: run A_load_data() first.
}
