% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_load_and_plot_all.R
\name{C_load_and_plot_all}
\alias{C_load_and_plot_all}
\title{Helper function for ease of calling from CLI}
\usage{
C_load_and_plot_all(input_json, input_tree, output_folder)
}
\arguments{
\item{input_json}{Argument for text filepath to metadata .json file name (SACCHARIS output).}

\item{input_tree}{Argument for text filepath to .tree file name (SACCHARIS output).}

\item{output_folder}{Argument for text folderpath output CAZY_TABLE_FINAL_GHxx_characterized_DATE.xlsx to.}
}
\description{
Generate circular plots of trees generated by SACCAHRIS v2. REQUIRED: Input .json and .tree files generated by SACCHARIS 2 and output folder
}
